requirements.txt
asgiref==3.3.4
astpretty==2.1.0
attrs==21.2.0
backcall==0.2.0
cognitive-complexity==1.2.0
decorator==5.0.9
Django==3.2.2
django-crispy-forms==1.11.2
eradicate==2.0.0
flake8==3.9.2
flake8-blind-except==0.2.0
flake8-builtins==1.5.3
flake8-cognitive-complexity==0.1.0
flake8-commas==2.0.0
flake8-comprehensions==3.5.0
flake8-django==1.1.2
flake8-eradicate==1.0.0
flake8-expression-complexity==0.0.9
flake8-logging-format==0.6.0
flake8-polyfill==1.0.2
flake8-print==4.0.0
flake8-variables-names==0.0.4
gunicorn==20.1.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
mccabe==0.6.1
parso==0.8.2
pep8-naming==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
whitenoise==5.2.0
